Add more solutions and minor corrections
Added solutions for problem 71, 72, 73, 74 and 75 in C and python
and made a few corrections in the code for other problems.

+++ C/p062.c
@@ -85,7 +85,7 @@ int count_digits(long int a)
    return count;
 }
 
-int is_permutation(int a, int b)
+int is_permutation(long int a, long int b)
 {
    int i;
    int digits1[10] = {0}, digits2[10] = {0};
